Update patch set 3
Patch Set 3:

Well this is needed if we enable multiple IDP and in that case it fix problems with collision we can have. I think that's great. In case of differentiate strategy and a user being registered on Gerrit, then twice or more but having the same email registered on # idp what will happen ?

Patch-set: 3
Label: Code-Review=0