Update patch set 2
Patch Set 2:

(2 comments)

Thanks Matthieu for the review <3. See my response and yes it could have been external but as long as cauth is not hosted outside of the SF umbrella I though it was useless to complexify. More work must come after sf-config and repoxplorer side to fully activate the new feature.

Patch-set: 2