[PoC] service-specific data injection in auth tkt
Add a mechanism allowing to inject specific data in the auth ticket.
The data is expected to be fetched at bind time with the external
service, and returned in the form of a dictionary.

Add jwt injection when zuul is a declared service. Please note that
due to what the resources backend stores the jwt's scope is limited
to repositories under the main gerrit connection (ie this won't
work for github or third party gerrits). JWT generation will have
to be handled upstream.

TG-1538

Change-Id: I8eaa95922b4b7c772a0929272c193ae3cb9cc9bc
Depends-On: I4dadaa09dbfb50e10bbfd7944194782e617ae4db
Depends-On: Ic0fa25682feb1eea51a4bbe8a9ad5d4223038879
Depends-On: Icd5e113218c67197b40b22dba423b2e750cb03ec