Logging: add a transaction ID, sanitize messages
A transaction ID is computed and added at the beginning of log lines
and on error messages to help with debugging.
User-facing error messages are also simplified to avoid potential
leaks.
Some logging messages are demoted from ERROR to DEBUG where it makes
more sense to do so.

Removed stevedor cap as it conflicted with pbr.

TG-1772

Co-Authored-By: Tristan De Cacqueray <tdecacq@redhat.com>
Change-Id: I9af472d1b67c88cf337676f4ea5cb763944302c2