Update patch set 4
Patch Set 4:

> (1 comment)
 > 
 > I don't see any transaction headers in cauth.log in CI runs. Do I
 > miss something ?

No, I consider that normal operations events (authenticated users are logged in, users failing to authenticate for non-cauth-related reasons are not logged in) should be logged as debug or less. You get transaction infos when there is a problem or at debug level. If it's not enough let me know!

Also I fixed the warning level in the repoxplorer service in the last patch.

Patch-set: 4