Update patch set 7
Patch Set 7:

Last PS7 significantly rewrote the propose logic to ensure the transaction ID is set properly, all the way to service plugins. To simplify the code I added a TransactionLogger and refactored logging module usage... It seems to be working as expected, but we should carefuly review how errors are handled.

Patch-set: 7