Update patch set 8
Patch Set 8:

(5 comments)

Very nice, thanks!!

Do the transaction IDs appear in the Exception messages? Because that's what gets sent back to the frontend when there is an error. This way a user can communicate a transaction ID that the operator can look for in logs. I haven't seen it done but I might have overlooked.

Patch-set: 8
Label: Code-Review=0