Update patch set 8
Patch Set 8:

Python3-devel needed. On nodes we install python-devel that is python2-devel on f29 and f30. So trying on f31 as python-devel is python3-devel. But tests are skipped I guess as on f31 with have py37 and not py36 ...
So now will try with https://softwarefactory-project.io/r/#/c/16956/ that should provide py2 and py3 devel on each f29/30/31.

Patch-set: 8
Label: Code-Review=0